requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.40.3
botocore==1.40.3
bs4==0.0.2
cachetools==5.5.2
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distro==1.9.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gensim==4.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
google-genai
google-adk
graphviz==0.21
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
imblearn==0.0
iterative-stratification==0.1.9
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-aws==0.2.30
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.8
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.12
lightgbm==4.6.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.6.3
multimethod==1.12
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.0.1
nbformat==5.10.4
networkx==3.5
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.99.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas 
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
propcache==0.3.2
protobuf>=3.20.3,<5.0.0
psutil==7.0.0
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil 
python-dotenv==1.1.1
pytz 
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.6.1
scikit-multilearn==0.2.0
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.0.0
six 
sklearn-compat==0.1.3
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
triton==3.3.1
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata 
uritemplate==4.2.0
urllib3==2.5.0
visions==0.8.1
wcwidth==0.2.13
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.3
xxhash==3.5.0
yarl==1.20.1
ydata-profiling==4.16.1
zstandard==0.23.0
optuna
hyperopt
tensorflow
huggingface_hub>=0.23
transformers[sentencepiece]>=4.41     
tokenizers>=0.19
safetensors>=0.4.0
protobuf
tf-keras
